requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==4.1.3
django-bootstrap-v5==1.0.11
django-bootstrap5==22.2
django-model-utils==4.3.1
djangorestframework==3.14.0
pytz==2022.6
soupsieve==2.3.2.post1
sqlparse==0.4.3
